+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 <!-- see keepachangelog.com for format ideas -->
 
+## 0.2.5 (2023-03-14)
+
+### Added
+
+- `TlsServerEngine` to handle TLS wrapping on any server-side TCP
+  stream.  Uses `rustls` and requires that the `rustls` feature is
+  enabled.
+
+- `TcpStreamBuf` improvements:
+  - `Read` and `Write` implementations
+  - Input EOF flag
+  - Ability to pause flush (for Windows before "write ready")
+  - Forward data from one `TcpStreamBuf` to another (`exchange`)
+
 ## 0.2.4 (2022-09-06)
 
 Update to `mio` version 0.8.*